Extract out new module `TokenMetadata`.
