+++ lib/wallet/api/http/Cardano/Wallet/Api/Types.hs
@@ -647,6 +647,7 @@ import qualified Cardano.Crypto.Wallet as CC
 import qualified Cardano.Wallet.Address.Derivation as AD
 import qualified Cardano.Wallet.Primitive.Types as W
 import qualified Cardano.Wallet.Primitive.Types.TokenFingerprint as W
+import qualified Cardano.Wallet.Primitive.Types.TokenMetadata as W
 import qualified Cardano.Wallet.Primitive.Types.TokenName as W
 import qualified Cardano.Wallet.Primitive.Types.TokenPolicy as W
 import qualified Codec.Binary.Bech32 as Bech32
