+++ lib/wallet/test/unit/Cardano/Wallet/TokenMetadataSpec.hs
@@ -15,15 +15,17 @@ import Cardano.Wallet.Primitive.Types.Hash
 import Cardano.Wallet.Primitive.Types.TokenMap
     ( AssetId (..)
     )
-import Cardano.Wallet.Primitive.Types.TokenName
-    ( nullTokenName
-    )
-import Cardano.Wallet.Primitive.Types.TokenPolicy
+import Cardano.Wallet.Primitive.Types.TokenMetadata
     ( AssetDecimals (..)
     , AssetLogo (..)
     , AssetMetadata (..)
     , AssetURL (..)
-    , TokenPolicyId (..)
+    )
+import Cardano.Wallet.Primitive.Types.TokenName
+    ( nullTokenName
+    )
+import Cardano.Wallet.Primitive.Types.TokenPolicy
+    ( TokenPolicyId (..)
     )
 import Cardano.Wallet.TokenMetadata
     ( BatchResponse (..)
